// src/types/Notificacao.ts
export type Notificacao = {
  id: string;
  form_efetivacao_seguro_fianca_tb?: string;
  form_quadro_anotacao?: string;
  form_seguro_fianca_empresarial_mais_2_anos?: string;
  form_seguro_fianca_empresarial_menos_2_anos?: string;
  form_seguro_fianca_residencial?: string;
  form_seguro_incendio?: string;
  form_seguro_incendio_comercial?: string;
  form_titulo_capitalizacao?: string;
};